Amend naming. Move Home Screen Widget into home module
