+++ app/src/test/kotlin/uk/govuk/app/data/local/AppDataStoreTest.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
-import uk.govuk.app.HomeScreenWidget
+import uk.govuk.app.home.HomeScreenWidget
 
 class AppDataStoreTest {
